Criando mapeamentos e organizando [oneToMany]

+++ src/main/java/com/urso/chat/repository/ChatRepository.java
@@ -3,5 +3,5 @@ package com.urso.chat.repository;
 import com.urso.chat.entity.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRepository extends JpaRepository<Chat,Integer> {
+public interface ChatRepository extends JpaRepository<Chat,Long> {
 }
